requirements.txt
python-telegram-bot[job-queue]==20.0
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3